set up the server and connection to MongoDB
