+++ server.js
@@ -1,8 +1,31 @@
+const express = require("express");
+const logger = require("morgan");
+const cors = require("cors");
 const mongoose = require("mongoose");
 
-const app = require("./app");
+require("dotenv").config();
+
+const app = express();
 
 const {DB_HOST, PORT = 3000} = process.env;
+
+const formatsLogger = app.get("env") === "development" ? "dev" : "short";
+const contactsRouter = require("./routes/api/contacts");
+
+app.use(logger(formatsLogger));
+app.use(cors());
+app.use(express.json());
+
+app.use("/api/contacts", contactsRouter);
+
+app.use((req, res) => {
+  res.status(404).json({message: "Not found"});
+});
+
+app.use((err, req, res, next) => {
+  const {status = 500, message = "Server error"} = err;
+  res.status(status).json({message});
+});
 
 mongoose
   .connect(DB_HOST)
